+++ packages/ui/src/components/InviteSignupBoard/InviteSignupBoard.tsx
@@ -1,9 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
 import {yupResolver} from "@hookform/resolvers";
+import Avatar from "@material-ui/core/Avatar";
 import Box from "@material-ui/core/Box";
 import Button from "@material-ui/core/Button";
 import Checkbox from "@material-ui/core/Checkbox";
+import Container from "@material-ui/core/Container";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Grid from "@material-ui/core/Grid";
 import TextField from "@material-ui/core/TextField";
@@ -15,38 +17,38 @@ import {useRouter} from "next/router";
 import * as React from "react";
 import "firebase/auth";
 import {useForm, Controller} from "react-hook-form";
-
 import * as Yup from "yup";
 
 import signinWithGoogle from "@sentrei/common/services/signinWithGoogle";
 import signup from "@sentrei/common/services/signup";
+import Invite from "@sentrei/types/models/Invite";
 
-import FormSection from "@sentrei/ui/components/FormSection";
 import MuiLink from "@sentrei/ui/components/MuiLink";
 
 import useBackdrop from "@sentrei/ui/hooks/useBackdrop";
 import useSnackbar from "@sentrei/ui/hooks/useSnackbar";
 
-import InviteFormSignupStyles from "./InviteFormSignupStyles";
+import InviteSignupBoardStyles from "./InviteSignupBoardStyles";
 
 export interface Props {
-  inviteId: string;
+  invite: Invite.Get;
+  namespaceId: string;
   spaceId: string;
 }
 
-export default function InviteFormSignup({
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  inviteId,
+export default function InviteSignupBoard({
+  invite,
+  namespaceId,
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   spaceId,
 }: Props): JSX.Element {
-  const classes = InviteFormSignupStyles();
+  const classes = InviteSignupBoardStyles();
   const {backdrop} = useBackdrop();
   const {snackbar} = useSnackbar();
   const {t, lang} = useTranslation();
   const {query} = useRouter();
 
-  const InviteFormSignupSchema = Yup.object().shape({
+  const InviteSignupBoardSchema = Yup.object().shape({
     email: Yup.string()
       .required(t("form:email.required"))
       .email(t("form.email.valid")),
@@ -55,7 +57,7 @@ export default function InviteFormSignup({
 
   const {control, register, errors, handleSubmit} = useForm({
     reValidateMode: "onChange",
-    resolver: yupResolver(InviteFormSignupSchema),
+    resolver: yupResolver(InviteSignupBoardSchema),
   });
 
   const google = (): void => {
@@ -90,12 +92,37 @@ export default function InviteFormSignup({
   };
 
   return (
-    <>
-      <FormSection
-        icon={<AccountCircleOutlinedIcon />}
-        title={t("auth:invite.title")}
-        size="sm"
-      />
+    <Container maxWidth="sm">
+      <Container component="main" maxWidth="md">
+        <div className={classes.paper}>
+          <Box p={3}>
+            <Avatar className={classes.avatar}>
+              <AccountCircleOutlinedIcon />
+            </Avatar>
+          </Box>
+          <Typography component="h1" variant="h3" gutterBottom>
+            {t("common:common.invite")}
+          </Typography>
+          {invite.email && (
+            <Typography component="h1" variant="h3" gutterBottom>
+              {invite.email}
+            </Typography>
+          )}
+          <Typography align="center" variant="h6" component="h6">
+            {t("common:common.inviteTo")} {namespaceId}
+          </Typography>
+          <Typography
+            align="center"
+            variant="body2"
+            color="textSecondary"
+            component="p"
+            gutterBottom
+          >
+            {t("common:common.inviteFrom")}
+            {invite.createdBy.name}
+          </Typography>
+        </div>
+      </Container>
       <Grid container spacing={3}>
         <Box p={1} />
         <Button
@@ -191,6 +218,6 @@ export default function InviteFormSignup({
           </Grid>
         </Grid>
       </Grid>
-    </>
+    </Container>
   );
 }