+++ packages/ui/src/components/RoomCard/RoomCardStyles.ts
@@ -1,4 +1,4 @@
-import {createStyles, makeStyles, Theme} from "@material-ui/core/styles";
+import {createStyles, fade, makeStyles, Theme} from "@material-ui/core/styles";
 
 const RoomScreenStyles = makeStyles((theme: Theme) =>
   createStyles({
@@ -10,6 +10,12 @@ const RoomScreenStyles = makeStyles((theme: Theme) =>
       maxWidth: "100%",
       height: theme.spacing(30),
       maxHeight: theme.spacing(30),
+    },
+    placeholder: {
+      backgroundColor: fade(
+        theme.palette.text.primary,
+        theme.palette.type === "light" ? 0.11 : 0.13,
+      ),
     },
     root: {
       boxShadow: "0 8px 40px -12px rgba(0,0,0,0.3)",